requirements/common.txt
django==5.0.1
djangorestframework==3.14.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1